+++ src/integrations/terminal/TerminalManager.ts
@@ -287,57 +287,4 @@ export class TerminalManager {
 		this.disposables.forEach((disposable) => disposable.dispose())
 		this.disposables = []
 	}
-
-	/**
-	 * Gets the terminal contents based on the number of commands to include
-	 * @param commands Number of previous commands to include (-1 for all)
-	 * @returns The selected terminal contents
-	 */
-	public async getTerminalContents(commands = -1): Promise<string> {
-		// Save current clipboard content
-		const tempCopyBuffer = await vscode.env.clipboard.readText()
-
-		try {
-			// Select terminal content
-			if (commands < 0) {
-				await vscode.commands.executeCommand("workbench.action.terminal.selectAll")
-			} else {
-				for (let i = 0; i < commands; i++) {
-					await vscode.commands.executeCommand("workbench.action.terminal.selectToPreviousCommand")
-				}
-			}
-
-			// Copy selection and clear it
-			await vscode.commands.executeCommand("workbench.action.terminal.copySelection")
-			await vscode.commands.executeCommand("workbench.action.terminal.clearSelection")
-
-			// Get copied content
-			let terminalContents = (await vscode.env.clipboard.readText()).trim()
-
-			// Restore original clipboard content
-			await vscode.env.clipboard.writeText(tempCopyBuffer)
-
-			if (tempCopyBuffer === terminalContents) {
-				// No terminal content was copied
-				return ""
-			}
-
-			// Process multi-line content
-			const lines = terminalContents.split("\n")
-			const lastLine = lines.pop()?.trim()
-			if (lastLine) {
-				let i = lines.length - 1
-				while (i >= 0 && !lines[i].trim().startsWith(lastLine)) {
-					i--
-				}
-				terminalContents = lines.slice(Math.max(i, 0)).join("\n")
-			}
-
-			return terminalContents
-		} catch (error) {
-			// Ensure clipboard is restored even if an error occurs
-			await vscode.env.clipboard.writeText(tempCopyBuffer)
-			throw error
-		}
-	}
 }
